pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sam3-prelabel-pipeline"
version = "0.1.0"
description = "Configurable SAM-3 image pre-labeling pipeline with COCO export."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=2,<2.4",
  "opencv-python>=4.8.1.78,<=4.10.0.84",
  "Pillow>=11,<12",
  "PyYAML",
  "tqdm",
  "inference-sdk",
]

[project.scripts]
sam3-prelabel = "sam3_prelabel.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
